moving test to tests.rc
