+++ ironfish-rust/src/transaction/mod.rs
@@ -71,6 +71,9 @@ mod utils;
 mod value_balances;
 mod version;
 
+#[cfg(test)]
+mod tests;
+
 pub use version::TransactionVersion;
 
 const SIGNATURE_HASH_PERSONALIZATION: &[u8; 8] = b"IFsighsh";
@@ -984,37 +987,4 @@ pub fn split_secret(
     }
 
     Ok((key_packages, pubkeys))
-}
-
-#[cfg(test)]
-mod tests {
-    use ironfish_frost::frost::{frost::keys::reconstruct, JubjubBlake2b512};
-
-    use super::*;
-
-    #[test]
-    fn test_split_secret() {
-        let mut rng = rand::thread_rng();
-
-        let key = SaplingKey::generate_key().spend_authorizing_key.to_bytes();
-
-        let config = SecretShareConfig {
-            min_signers: 2,
-            max_signers: 3,
-            secret: key.to_vec(),
-        };
-
-        let (key_packages, _) =
-            split_secret(&config, frost::keys::IdentifierList::Default, &mut rng).unwrap();
-        assert_eq!(key_packages.len(), 3);
-
-        let key_parts: Vec<_> = key_packages.values().cloned().collect();
-
-        let signing_key =
-            reconstruct::<JubjubBlake2b512>(&key_parts).expect("key reconstruction failed");
-
-        let scalar = signing_key.to_scalar();
-
-        assert_eq!(scalar.to_bytes(), key);
-    }
 }
